+++ CHANGELOG.md
@@ -10,6 +10,11 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 ### Fixed
 * `ProgIter` now correctly checks if it needs to displays a message on every iteration.
 
+### Added
+* `ub.repr2` now accept type name strings at register time (which makes it
+  easier to lazy-load heavy libraries)
+* `ub.repr2` now handles torch.Tensor and pandas.DataFrame objects by default
+
 
 ### Changed
 * `ProgIter.step` now respects update freq, and will not update the estimates